+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.0
+
+* New design
+* Improved the `@requires` annotation to support external vendors, and custom URL
+* Added a search engine to the generated documentation
+* Fixed an issue with `@link` not working correctly
+* Added `examples` to `.gitignore`
+
 ## 1.0.2
 
 * Fixed an issue with config path resolving to false